Add session heartbeat handler
This adds a new handler and a plugin for JLab, which periodically checks
the logged-in user's Chameleon session. Essentially, it checks that their
access_token could be refreshed. If they cannot, it means their SSO session
timed out on the backend and they need to be logged out.

However, we can attempt to reauthenticate them transparently. It tries to
do that by taking them to a special Hub handler that triggers a frontend-only
logout and then goes through the IdP login transaction again. If the user
is already logged in to the IdP, this will just take them back to where they
were, effectively.