Fixup various issues with SSH provisioner
The main big issue was that paramiko has no great way to start
a login shell, and users may very well have a custom environment
that needs sourced in order to find things. The most predictable
thing to do is try to get a login shell, but it turns out this
is really arduous to do, especially when you want to capture stdout
and exit status of a particular process.