Add artifact link widget and abandon sqlite3 database
This makes it so artifacts loaded by searching for .trovi.json files in
the users working dir (and from the database for legacy purposes). Newly
created artifacts create these files instead of inserting into the
database. There is an additional "Link" widget, that allows you to link
a directory to an existing trovi artifact, so you can upload a new
version and edit its metadata.